+++ RougeLike/Source/Utilities/Maps/PropertyMap.cpp
@@ -22,7 +22,7 @@ void PropertyMap::fill(const XMLNode& node)
 	for (StringMap::const_iterator iter = map.begin(); iter != map.end(); iter++)
 	{
 		PropertyType name(iter->first);
-		float value = atof(iter->second.c_str());
+		float value = (float)atof(iter->second.c_str());
 		mData[name] = value;
 	}
 }
